# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rakshak AI - Crime Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accused:
    get:
      operationId: listAccused
      tags: [accused]
      summary: List accused with risk scores
      parameters:
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Accused list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accused"
  /accused/{id}:
    get:
      operationId: getAccused
      tags: [accused]
      summary: Get offender profile with risk scoring
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offender profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccusedProfile"
        "404":
          description: Not found
  /accused/{id}/network:
    get:
      operationId: getAccusedNetwork
      tags: [accused]
      summary: Criminal network graph for a specific accused
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Network graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkGraph"

  # ─── Criminal Network ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Accused:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        district:
          type: string
        totalFirs:
          type: integer
        riskScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        isRepeatOffender:
          type: boolean
        crimeTypes:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - age
        - gender
        - district
        - totalFirs
        - riskScore
        - riskLevel
        - isRepeatOffender
        - crimeTypes
    AccusedProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
        district:
          type: string
        totalFirs:
          type: integer
        riskScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        isRepeatOffender:
          type: boolean
        hasViolentCrimes:
          type: boolean
        usedWeapons:
          type: boolean
        networkSize:
          type: integer
        crimeTypes:
          type: array
          items:
            type: string
        firHistory:
          type: array
          items:
            $ref: "#/components/schemas/FirReference"
        behavioralProfile:
          type: string
        riskFactors:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - age
        - gender
        - district
        - totalFirs
        - riskScore
        - riskLevel
        - isRepeatOffender
        - hasViolentCrimes
        - usedWeapons
        - networkSize
        - crimeTypes
        - firHistory
        - behavioralProfile
        - riskFactors
    FirReference:
      type: object
      properties:
        firId:
          type: integer
        firNumber:
          type: string
        crimeType:
          type: string
        date:
          type: string
        policeStation:
          type: string
        status:
          type: string
      required:
        - firId
        - firNumber
        - crimeType
        - date
        - policeStation
        - status

    # Network Graph
    NetworkGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NetworkNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/NetworkEdge"
        stats:
          $ref: "#/components/schemas/NetworkStats"
      required:
        - nodes
        - edges
        - stats
    NetworkNode:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [accused, victim, location, vehicle, phone, bank, fir]
        riskLevel:
          type: ["string", "null"]
        extraInfo:
          type: ["string", "null"]
      required:
        - id
        - label
        - type
    NetworkEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        label:
          type: string
        weight:
          type: number
      required:
        - id
        - source
        - target
        - label
        - weight
    NetworkStats:
      type: object
      properties:
        totalNodes:
          type: integer
        totalEdges:
          type: integer
        gangsDetected:
          type: integer
        centralNode:
          type: string
      required:
        - totalNodes
        - totalEdges
        - gangsDetected
        - centralNode

    # Analytics
...