# lib/api-zod/src/generated/types/networkNode.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Rakshak AI - Crime Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkNodeType } from './networkNodeType';

export interface NetworkNode {
  id: string;
  label: string;
  type: NetworkNodeType;
  /** @nullable */
  riskLevel?: string | null;
  /** @nullable */
  extraInfo?: string | null;
}